Medical: Add variable "avgstat" for WoundInfo
The problem with getting the average status with a function is that you
can't use the function if you're trying to use it from a UI scope.

Since copying the whole function just to put a UI scope on it isn't that
great of an idea, I've decided to just add a variable instead, and make
the function just update the variable.