Looting: Multi fallback loot with new RNG distrib
Fixes the multiple fallback loot behaviour (which e9cd9af regressed), using an
exponential falloff distribution in order to make larger amounts of loot rarer,
one which is fully configurable using CVars.

There is still a little bit of refactoring and cleanup left to do in the code,
and testing for edge cases with the RNG distribution configuration (as, since
xponentiation is used, some unhandled mathematical edge cases may result in
crashes or unexpected behaviour).