Looting: More balanced UaS_LootingFactor default
This changes the default from 3 to 2, which is more balanced.
(Pools with 4 and 6 were too common with 3.)

Please note that this looting factor is still tailored for this new
exponential falloff distribution based algorithm, and it doesn't have
anything to do with its old meaning anymore, where it only made sense as
a value between 0 and 1.